Conflicts:
	DCloudGateway/src/gateway/api.py
	DCloudGateway/src/gateway/api_restore_conf.py

+++ DCloudGateway/src/gateway/BackupToCloud.py
@@ -10,65 +10,72 @@ import shutil
 import grp
 import pwd
 import stat
+import common
 import simplejson as json
 from time import gmtime, strftime
 from SwiftClient import *
 
 TEMP_PATH = '/tmp/backuptocloud'
+log = common.getLogger(name="Backup To Cloud", conf="/etc/delta/Gateway.ini")
 
 class BackupToCloud():
     """
-	
     """
     def __init__(self, fileList = None, cloudObject = None):
         self._fileList = fileList
         self._cloudObject = cloudObject
         if self._fileList is None:
-            self._fileList = ['/etc/network/interfaces', '/home/hungic/.bashrc']
+            self._fileList = ['/etc/network/interfaces', '/home/hungic/.bashrc', '/tmp/aaa']
             print self._fileList
         if self._cloudObject is None:
             self._cloudObject = SwiftClient()
         self._metaData = {}
     
-    def backupToCloud(self):
+    def backup(self):
         """
         """
         self._datetime = strftime("%Y%m%d%H%M", gmtime())
+        self._tarFileName = '%s_gw_conf_backup.tar.gz' % self._datetime
         self.copyFile()
-        #self.tarFile()
-        #self.sendToCloud()
+        self.tarFile()
+        self.sendToCloud()
 
     def copyFile(self):
         """
         """
-        shutil.rmtree(TEMP_PATH)
+        if os.path.exists(TEMP_PATH):
+            shutil.rmtree(TEMP_PATH)
         os.mkdir(TEMP_PATH)
         i = 1
         for file in self._fileList:
-            statInfo = os.stat(file)
-            user = pwd.getpwuid(statInfo.st_uid)[0]
-            group = grp.getgrgid(statInfo.st_gid)[0]
-            fileMode = stat.S_IMODE(statInfo.st_mode)
-            print user, group, fileMode
-            key = ''.join(['file', str(i)])
-            self._metaData[key] = {'fname': os.path.basename(file), 'fpath': os.path.dirname(file), 'user': user, 'group': group, 'chmod': fileMode}
-            shutil.copy(file, TEMP_PATH)
-            i = i + 1
+            if os.path.exists(file):
+                statInfo = os.stat(file)
+                user = pwd.getpwuid(statInfo.st_uid)[0]
+                group = grp.getgrgid(statInfo.st_gid)[0]
+                fileMode = stat.S_IMODE(statInfo.st_mode)
+                print user, group, fileMode
+                key = ''.join(['file', str(i)])
+                self._metaData[key] = {'fname': os.path.basename(file), 'fpath': os.path.dirname(file), 'user': user, 'group': group, 'chmod': fileMode}
+                shutil.copy2(file, TEMP_PATH)
+                i = i + 1
+            else:
+                log.warning('File(%s) is not exist' % file)
         print self._metaData
         try:
-            print ''.join([TEMP_PATH, '/metadata.txt'])
-            fd = open(''.join([TEMP_PATH, '/metadata.txt']), 'w+')
-            fd.write(json.dumps(self._metaData))
-            fd.close()
+            with open(''.join([TEMP_PATH, '/metadata.txt']), 'w+') as fd:
+                fd.write(json.dumps(self._metaData))
         except IOError as e:
-            print e
+            raise CreateMetaDataError()
 
     def tarFile(self):
+        os.system('tar -zcvf %s %s' % (self._tarFileName, TEMP_PATH))
         
+    def sendToCloud(self):
+        self._cloudObject.upload('config', self._tarFileName)
 
 def main(argv = None):
     test = BackupToCloud()
-    test.backupToCloud()
+    test.backup()
 if __name__ == '__main__':
     main()
-		
+
